Move client/tlsclient to pkg/tlsclient
The client directory is used for packages for clients of the stack, but
tlsclient is used by the stack for talking to couchdb, and as such,
should be in pkg and not client.
